Renamed segments package
