+++ src/main/java/io/github/programminglife2016/pl1_2016/Launcher.java
@@ -1,7 +1,7 @@
 package io.github.programminglife2016.pl1_2016;
 
 import io.github.programminglife2016.pl1_2016.parser.JsonSerializable;
-import io.github.programminglife2016.pl1_2016.parser.segments.SegmentParser;
+import io.github.programminglife2016.pl1_2016.parser.nodes.SegmentParser;
 import io.github.programminglife2016.pl1_2016.server.api.RestServer;
 import io.github.programminglife2016.pl1_2016.server.Server;
