+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/JsonSerializable.java
@@ -2,9 +2,9 @@ package io.github.programminglife2016.pl1_2016.parser;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import io.github.programminglife2016.pl1_2016.parser.segments.NodeCollectionSerializer;
-import io.github.programminglife2016.pl1_2016.parser.segments.NodeList;
-import io.github.programminglife2016.pl1_2016.parser.segments.NodeMap;
+import io.github.programminglife2016.pl1_2016.parser.nodes.NodeCollectionSerializer;
+import io.github.programminglife2016.pl1_2016.parser.nodes.NodeList;
+import io.github.programminglife2016.pl1_2016.parser.nodes.NodeMap;
 
 /**
  * Data is sent by the server in JSON format. Classes extending this method can convert their
