+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/nodes/NodeList.java
@@ -1,4 +1,4 @@
-package io.github.programminglife2016.pl1_2016.parser.segments;
+package io.github.programminglife2016.pl1_2016.parser.nodes;
 
 import java.util.Arrays;
 import java.util.Collection;
