+++ src/main/java/io/github/programminglife2016/pl1_2016/parser/nodes/PositionManager.java
@@ -1,4 +1,4 @@
-package io.github.programminglife2016.pl1_2016.parser.segments;
+package io.github.programminglife2016.pl1_2016.parser.nodes;
 
 /**
  * Manager for calculating postions of nodes. Classes implementing this interface
