+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/SegmentMapTest.java
@@ -2,7 +2,7 @@
 
 package io.github.programminglife2016.pl1_2016.parser;
 
-import io.github.programminglife2016.pl1_2016.parser.segments.NodeMap;
+import io.github.programminglife2016.pl1_2016.parser.nodes.NodeMap;
 import org.junit.Before;
 
 /**
